Added PCR closing in camp list
