+++ lib/common-lib/src/services/campService.js
@@ -906,3 +906,26 @@ export const closePcrCamp = async (params = {}, header = {}) => {
     return handleResponseException(e)
   }
 }
+export const multipleClosePcr = async (params = {}, header = {}) => {
+  let headers = {
+    ...header,
+    'Content-Type': 'application/json',
+    Authorization: 'Bearer ' + localStorage.getItem('token')
+  }
+  try {
+    const result = await post(
+      `${process.env.REACT_APP_API_URL}/camp/admin/multiple_end_pcr`,
+      params,
+      {
+        headers: headers || {}
+      }
+    )
+    if (result?.data) {
+      return result?.data
+    } else {
+      return {}
+    }
+  } catch (e) {
+    return handleResponseException(e)
+  }
+}
